requirements/base.txt
# Django
# ------------------------------------------------------------------------------
django==4.1.7
django-environ==0.10.0          # configure your Django application with environment variables
django-admin-interface==0.24.2  # a modern responsive flat admin interface customizable by the admin itself

# Django Rest Framework dependencies
# ------------------------------------------------------------------------------
djangorestframework==3.14.0     # Web APIs for Django
markdown==3.3.7                 # Markdown support for Python (for the browsable API.)
django-filter==23.1             # A reusable Django application for allowing users to filter querysets dynamically
